+++ CHANGELOG.md
@@ -1,5 +1,10 @@
+### 5.0.0
+* [BREAKING ENHANCEMENT] Follow babeljs major semver
+* [BREAKING ENHANCEMENT] changed `experimental` to `stage` which now accepts an integer. `stage: 0` is the same as `experimental: true`
+* [FEATURE] added options for `ignore`, `only` and `extensions`.
+
 ### 2.1.0
 
 * [ENHANCEMENT] Bumped Babel version.
 * [ENHANCEMENT] Made `experimental` flag configurable.
-* [ENHANCEMENT] Made `losse` flage configurable.
+* [ENHANCEMENT] Made `loose` flag configurable.